create PizzaSelectItem component which generates each individual pizza menu item
